drill-recommender/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
# firebase-admin==6.2.0  # Optional - only needed if using Firebase authentication
gunicorn==21.2.0

